[api refactor] Begin process of renaming `LogStream` to `Logger`.
