+++ lib/winston.js
@@ -35,7 +35,7 @@ winston.format    = require('./winston/formats');
 //
 winston.ExceptionHandler = require('./winston/exception-handler');
 winston.Container        = require('./winston/container').Container;
-winston.Logger           = require('./winston/logger').Logger;
+winston.Logger           = require('./winston/log-stream');
 winston.LogStream        = require('./winston/log-stream');
 winston.Transport        = require('winston-transport');
