+++ lib/winston/common.js
@@ -60,80 +60,6 @@ exports.setLevels = function (target, past, current) {
 };
 
 //
-// ### function timestamp ()
-// Returns a timestamp string for the current time.
-//
-exports.timestamp = function () {
-  return new Date().toISOString();
-};
-
-//
-// ### function serialize (obj, key)
-// #### @obj {Object|literal} Object to serialize
-// #### @key {string} **Optional** Optional key represented by obj in a larger object
-// Performs simple comma-separated, `key=value` serialization for Loggly when
-// logging to non-JSON inputs.
-//
-exports.serialize = function (obj, key) {
-  // symbols cannot be directly casted to strings
-  if (typeof key === 'symbol') {
-    key = key.toString()
-  }
-  if (typeof obj === 'symbol') {
-    obj = obj.toString()
-  }
-
-  if (obj === null) {
-    obj = 'null';
-  }
-  else if (obj === undefined) {
-    obj = 'undefined';
-  }
-  else if (obj === false) {
-    obj = 'false';
-  }
-
-  if (typeof obj !== 'object') {
-    return key ? key + '=' + obj : obj;
-  }
-
-  if (obj instanceof Buffer) {
-    return key ? key + '=' + obj.toString('base64') : obj.toString('base64');
-  }
-
-  var msg = '',
-      keys = Object.keys(obj),
-      length = keys.length;
-
-  for (var i = 0; i < length; i++) {
-    if (Array.isArray(obj[keys[i]])) {
-      msg += keys[i] + '=[';
-
-      for (var j = 0, l = obj[keys[i]].length; j < l; j++) {
-        msg += exports.serialize(obj[keys[i]][j]);
-        if (j < l - 1) {
-          msg += ', ';
-        }
-      }
-
-      msg += ']';
-    }
-    else if (obj[keys[i]] instanceof Date) {
-      msg += keys[i] + '=' + obj[keys[i]];
-    }
-    else {
-      msg += exports.serialize(obj[keys[i]], keys[i]);
-    }
-
-    if (i < length - 1) {
-      msg += ', ';
-    }
-  }
-
-  return msg;
-};
-
-//
 // ### function tailFile (options, callback)
 // #### @options {Object} Options for tail.
 // #### @callback {function} Callback to execute on every line.
