+++ lib/winston/log-stream.js
@@ -3,12 +3,15 @@
 var stream = require('stream'),
     util = require('util'),
     format = util.format,
-    ExceptionHandler = require('./exception-handler'),
-    Profiler = require('./profiler'),
-    TransportStream = require('winston-transport'),
-    LegacyTransportStream = require('winston-transport/legacy'),
+    async = require('async'),
     isStream = require('isstream'),
+    TransportStream = require('winston-transport'),
+    ExceptionHandler = require('./exception-handler'),
+    LegacyTransportStream = require('winston-transport/legacy'),
+    Profiler = require('./profiler'),
     common = require('./common');
+
+const formatRegExp = /%[sdj%]/g;
 
 /*
  * Constructor function for the Logger object responsible
@@ -285,6 +288,122 @@ LogStream.prototype.close = function () {
 LogStream.prototype.setLevels = function (target) {
   common.setLevels(this, this.levels, target);
   return this;
+};
+
+//
+// ### function query (options, callback)
+// #### @options {Object} Query options for this instance.
+// #### @callback {function} Continuation to respond to when complete.
+// Queries the all transports for this instance with the specified `options`.
+// This will aggregate each transport's results into one object containing
+// a property per transport.
+//
+LogStream.prototype.query = function (options, callback) {
+  if (typeof options === 'function') {
+    callback = options;
+    options = {};
+  }
+
+  var self = this,
+      options = options || {},
+      results = {},
+      query = common.clone(options.query) || {},
+      transports;
+
+  //
+  // Helper function to query a single transport
+  //
+  function queryTransport(transport, next) {
+    if (options.query) {
+      options.query = transport.formatQuery(query);
+    }
+
+    transport.query(options, function (err, results) {
+      if (err) {
+        return next(err);
+      }
+
+      next(null, transport.formatResults(results, options.format));
+    });
+  }
+
+  //
+  // Helper function to accumulate the results from
+  // `queryTransport` into the `results`.
+  //
+  function addResults(transport, next) {
+    queryTransport(transport, function (err, result) {
+      //
+      // queryTransport could potentially invoke the callback
+      // multiple times since Transport code can be unpredictable.
+      //
+      if (next) {
+        result = err || result;
+        if (result) {
+          results[transport.name] = result;
+        }
+
+        next();
+      }
+
+      next = null;
+    });
+  }
+
+  //
+  // Iterate over the transports in parallel setting the
+  // appropriate key in the `results`
+  //
+  async.forEach(this.transports.filter(function (transport) {
+    return !!transport.query;
+  }), addResults, function () {
+    callback(null, results);
+  });
+};
+
+//
+// ### function stream (options)
+// #### @options {Object} Stream options for this instance.
+// Returns a log stream for all transports. Options object is optional.
+//
+LogStream.prototype.stream = function (options) {
+  options = options || {};
+
+  var self = this,
+      out = new stream.Stream(),
+      streams = [];
+
+  out._streams = streams;
+  out.destroy = function () {
+    var i = streams.length;
+    while (i--) streams[i].destroy();
+  };
+
+  //
+  // Create a list of all transports for this instance.
+  //
+  this.transports.filter(function (transport) {
+    return !!transport.stream;
+  }).forEach(function (transport) {
+    var stream = transport.stream(options);
+    if (!stream) return;
+
+    streams.push(stream);
+
+    stream.on('log', function (log) {
+      log.transport = log.transport || [];
+      log.transport.push(transport.name);
+      out.emit('log', log);
+    });
+
+    stream.on('error', function (err) {
+      err.transport = err.transport || [];
+      err.transport.push(transport.name);
+      out.emit('error', err);
+    });
+  });
+
+  return out;
 };
 
 //
