+++ lib/winston/logger.js
@@ -15,16 +15,16 @@ var events = require('events'),
 
 const formatRegExp = /%[sdj%]/g;
 
-//
-// ### function Logger (options)
-// #### @options {Object} Options for this instance.
-// Constructor function for the Logger object responsible
-// for persisting log messages and metadata to one or more transports.
-//
-var Logger = exports.Logger = function (options) {
-  // events.EventEmitter.call(this);
-  this.configure(options);
-};
+// //
+// // ### function Logger (options)
+// // #### @options {Object} Options for this instance.
+// // Constructor function for the Logger object responsible
+// // for persisting log messages and metadata to one or more transports.
+// //
+// var Logger = exports.Logger = function (options) {
+//   // events.EventEmitter.call(this);
+//   this.configure(options);
+// };
 
 //
 // Inherit from `events.EventEmitter`.
@@ -222,149 +222,149 @@ var Logger = exports.Logger = function (options) {
 //   return this;
 // };
 
+// //
+// // ### function query (options, callback)
+// // #### @options {Object} Query options for this instance.
+// // #### @callback {function} Continuation to respond to when complete.
+// // Queries the all transports for this instance with the specified `options`.
+// // This will aggregate each transport's results into one object containing
+// // a property per transport.
+// //
+// Logger.prototype.query = function (options, callback) {
+//   if (typeof options === 'function') {
+//     callback = options;
+//     options = {};
+//   }
 //
-// ### function query (options, callback)
-// #### @options {Object} Query options for this instance.
-// #### @callback {function} Continuation to respond to when complete.
-// Queries the all transports for this instance with the specified `options`.
-// This will aggregate each transport's results into one object containing
-// a property per transport.
+//   var self = this,
+//       options = options || {},
+//       results = {},
+//       query = common.clone(options.query) || {},
+//       transports;
 //
-Logger.prototype.query = function (options, callback) {
-  if (typeof options === 'function') {
-    callback = options;
-    options = {};
-  }
-
-  var self = this,
-      options = options || {},
-      results = {},
-      query = common.clone(options.query) || {},
-      transports;
-
-  //
-  // Helper function to query a single transport
-  //
-  function queryTransport(transport, next) {
-    if (options.query) {
-      options.query = transport.formatQuery(query);
-    }
-
-    transport.query(options, function (err, results) {
-      if (err) {
-        return next(err);
-      }
-
-      next(null, transport.formatResults(results, options.format));
-    });
-  }
-
-  //
-  // Helper function to accumulate the results from
-  // `queryTransport` into the `results`.
-  //
-  function addResults(transport, next) {
-    queryTransport(transport, function (err, result) {
-      //
-      // queryTransport could potentially invoke the callback
-      // multiple times since Transport code can be unpredictable.
-      //
-      if (next) {
-        result = err || result;
-        if (result) {
-          results[transport.name] = result;
-        }
-
-        next();
-      }
-
-      next = null;
-    });
-  }
-
-  //
-  // If an explicit transport is being queried then
-  // respond with the results from only that transport
-  //
-  if (options.transport) {
-    options.transport = options.transport.toLowerCase();
-    return queryTransport(this.transports[options.transport], callback);
-  }
-
-  //
-  // Create a list of all transports for this instance.
-  //
-  transports = this._names.map(function (name) {
-    return self.transports[name];
-  }).filter(function (transport) {
-    return !!transport.query;
-  });
-
-  //
-  // Iterate over the transports in parallel setting the
-  // appropriate key in the `results`
-  //
-  async.forEach(transports, addResults, function () {
-    callback(null, results);
-  });
-};
-
+//   //
+//   // Helper function to query a single transport
+//   //
+//   function queryTransport(transport, next) {
+//     if (options.query) {
+//       options.query = transport.formatQuery(query);
+//     }
 //
-// ### function stream (options)
-// #### @options {Object} Stream options for this instance.
-// Returns a log stream for all transports. Options object is optional.
+//     transport.query(options, function (err, results) {
+//       if (err) {
+//         return next(err);
+//       }
 //
-Logger.prototype.stream = function (options) {
-  var self = this,
-      options = options || {},
-      out = new Stream,
-      streams = [],
-      transports;
+//       next(null, transport.formatResults(results, options.format));
+//     });
+//   }
+//
+//   //
+//   // Helper function to accumulate the results from
+//   // `queryTransport` into the `results`.
+//   //
+//   function addResults(transport, next) {
+//     queryTransport(transport, function (err, result) {
+//       //
+//       // queryTransport could potentially invoke the callback
+//       // multiple times since Transport code can be unpredictable.
+//       //
+//       if (next) {
+//         result = err || result;
+//         if (result) {
+//           results[transport.name] = result;
+//         }
+//
+//         next();
+//       }
+//
+//       next = null;
+//     });
+//   }
+//
+//   //
+//   // If an explicit transport is being queried then
+//   // respond with the results from only that transport
+//   //
+//   if (options.transport) {
+//     options.transport = options.transport.toLowerCase();
+//     return queryTransport(this.transports[options.transport], callback);
+//   }
+//
+//   //
+//   // Create a list of all transports for this instance.
+//   //
+//   transports = this._names.map(function (name) {
+//     return self.transports[name];
+//   }).filter(function (transport) {
+//     return !!transport.query;
+//   });
+//
+//   //
+//   // Iterate over the transports in parallel setting the
+//   // appropriate key in the `results`
+//   //
+//   async.forEach(transports, addResults, function () {
+//     callback(null, results);
+//   });
+// };
 
-  if (options.transport) {
-    var transport = this.transports[options.transport];
-    delete options.transport;
-    if (transport && transport.stream) {
-      return transport.stream(options);
-    }
-  }
-
-  out._streams = streams;
-  out.destroy = function () {
-    var i = streams.length;
-    while (i--) streams[i].destroy();
-  };
-
-  //
-  // Create a list of all transports for this instance.
-  //
-  transports = this._names.map(function (name) {
-    return self.transports[name];
-  }).filter(function (transport) {
-    return !!transport.stream;
-  });
-
-  transports.forEach(function (transport) {
-    var stream = transport.stream(options);
-    if (!stream) return;
-
-    streams.push(stream);
-
-    stream.on('log', function (log) {
-      log.transport = log.transport || [];
-      log.transport.push(transport.name);
-      out.emit('log', log);
-    });
-
-    stream.on('error', function (err) {
-      err.transport = err.transport || [];
-      err.transport.push(transport.name);
-      out.emit('error', err);
-    });
-  });
-
-  return out;
-};
+// //
+// // ### function stream (options)
+// // #### @options {Object} Stream options for this instance.
+// // Returns a log stream for all transports. Options object is optional.
+// //
+// Logger.prototype.stream = function (options) {
+//   var self = this,
+//       options = options || {},
+//       out = new Stream,
+//       streams = [],
+//       transports;
+//
+//   if (options.transport) {
+//     var transport = this.transports[options.transport];
+//     delete options.transport;
+//     if (transport && transport.stream) {
+//       return transport.stream(options);
+//     }
+//   }
+//
+//   out._streams = streams;
+//   out.destroy = function () {
+//     var i = streams.length;
+//     while (i--) streams[i].destroy();
+//   };
+//
+//   //
+//   // Create a list of all transports for this instance.
+//   //
+//   transports = this._names.map(function (name) {
+//     return self.transports[name];
+//   }).filter(function (transport) {
+//     return !!transport.stream;
+//   });
+//
+//   transports.forEach(function (transport) {
+//     var stream = transport.stream(options);
+//     if (!stream) return;
+//
+//     streams.push(stream);
+//
+//     stream.on('log', function (log) {
+//       log.transport = log.transport || [];
+//       log.transport.push(transport.name);
+//       out.emit('log', log);
+//     });
+//
+//     stream.on('error', function (err) {
+//       err.transport = err.transport || [];
+//       err.transport.push(transport.name);
+//       out.emit('error', err);
+//     });
+//   });
+//
+//   return out;
+// };
 
 // //
 // // ### function handleExceptions ([tr0, tr1...] || tr0, tr1, ...)
